requirements.txt
Django==2.2.6
psycopg2==2.8.3
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
django-environ==0.4.5
django-rest-swagger==2.1.2